Bugfix [#28967] Attempting to add a new pose to the Pose Library causes Blender 2.60 RC2 to crash
This commit just rolls back part of r.40868, which was causing crashes
when trying to treat id-property-groups as bpy.type.Property to access
a special "rna_type" attribute added as part of said commit.

The underlying Py-API voodoo here is far too evil (along with the
myriad of ways of creating custom props) to work out an API fix for,
but at least we don't get anymore crashes now. In tests here, this
check even seems redundant!


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@41171 954f8c5b-7b00-dc11-b283-0030488c597c