Follow-up commit to r.41765
Reviewed behaviour of selection operators, and decided that ultimately, it's
better if select left/right/column didn't change the channel selections at all.
This is because with the highlighting of the active curve nowadays, it's a bit
distracting to suddenly lose track of it after performing these operations, when
you may have been trying to select all of the keyframes on that curve for
further tweaking.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@41874 954f8c5b-7b00-dc11-b283-0030488c597c