Partial Bugfix 2: [#29229] Outliner RMB commands unexpected results
RenderLayers and RenderPasses don't show the generic Hide/Unhide/Select/Deselect
popup which is irrelevant for this use case. I've included a commented-out call
here that can be replaced when we have some operations which can be performed on
this data (*)

(*) For new devs looking to get into blender dev, this could be a nice little
project to work on.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@41900 954f8c5b-7b00-dc11-b283-0030488c597c