Patch #29336: renaming UV (Texture) Layer to UV Map in the user interface, by Gaia Clary.
Rationale: the name was confusing and not always used consistently, and this
map itself is not something that can be layered, rather the map can be used
as texture coordinates in some layered setup.

The original intent was to indicate this contained more than just UV's, but
the game engine settings have already been moved out, and apparently users
didn't really get this from the name anyway.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@42098 954f8c5b-7b00-dc11-b283-0030488c597c