Fix [#29410] Blender svn 2.60.5 r42177, compositor, glare node bug.
The problem was in fact in the changes made to blend_ramp() func at rev42164. The checks on green pointer (if(g) {...) were checking the existance of a valid value (i.e. if input color was rgb or only shades of grey in r pointer), but as we get now a full float[3] pointer, this test is no more possible. And doing if(r_col[1]) {... was wrong, as it only broke things in case green was 0.0 (which was the case with the generated glare color of bug report).

So, just removed all those tests!

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@42181 954f8c5b-7b00-dc11-b283-0030488c597c