Fix #30102: Vertex and edge incorrect selections
Select Linked operator was setting Limit by Seams property automatically
depending on current selection mode. But this property only used to be
set to truth if selection mode is set to Face, in other cases this property
is being reused from previous operator run. This leads to incorrect
behavior of selecting linked in vertex mode after this operator was
used in face selection mode.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@43979 954f8c5b-7b00-dc11-b283-0030488c597c