Fix for hardcoded to scons rules python abi flags and wrong behavior of python bundling on systems where python is installed to /usr/lib64
Now ABI flags are automatically detecting (by checking all available flags
and checking if there's include directory exists for flag).
Also, automatically set PYTHON_LIBPATH to /usr/lib64 if python
scripts are stored in this folder.
Bundling python on *nix platforms is now checks if python is installed
to lib64 directory and if it is, python will be bundled to lib64
folder instead of lib.

This will make building on openSUSE a bit less annoying

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@43992 954f8c5b-7b00-dc11-b283-0030488c597c