Fix #30406: Hooks ignore Automatic handles setting in curves
Update handles positions after applying modifiers which seems to be expected behavior.
The only currently unsolved issue is about updating aligned handles because this needs
to determine in which order handles need to be recalculated which currently depends on
selection flags and which is quite tricky to do when running modifiers and animation data,
so currently just not update their positions for now.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@44836 954f8c5b-7b00-dc11-b283-0030488c597c