Adds support for utf paths on Windows.
Not all file formats/calls are supported yet. It will be expended.

Please from now on use BLI_fopen, BLI_* for file manipulations.
For non-windows systems BLI_fopen just calls fopen.
For Windows, the utf-8 string is translated to utf-16 string in order to call UTF version of the function.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@45012 954f8c5b-7b00-dc11-b283-0030488c597c