Patch [#30255] Center Last Stroke
Enables the NUMPERIOD view centering operator to work in sculpt mode.  Hitting NUMPERIOD while in sculpt mode will center the view on the end of the last sculpting stroke made by the user.

This is useful for quickly refocusing on the current work area without fussing with the view controls.

It does not zoom into the stroke (slide only).

It does nothing if there have been no strokes.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@46218 954f8c5b-7b00-dc11-b283-0030488c597c