Patch [#30965] Cancel Sculpt Stroke w/ ESCAPE
If the RMB has not been released after starting a sculpt stroke, then hitting escape will cancel the stroke in progress and undo any changes to the mesh.

This is a slightly faster work-flow than using undo, is a feature available in other paint programs, and also puts in place the infrastructure to add other keys later that could tweak strokes in different ways.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@46219 954f8c5b-7b00-dc11-b283-0030488c597c