Fix #31350, by Sergey Sharybin.
This happens because of how output node index is initializing in assign_index function: itterator goes to the beginning of the nodes list using node->prev and then reviews the whole node list to find first unused index. The problem is that node's initialization now is getting called before node was added to node tree, so all output nodes have got equal index.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@46424 954f8c5b-7b00-dc11-b283-0030488c597c