Fix #31747, broken image sequence loading after r47432.
The image file name function was updating the iuser->framenr using a supposed cfra parameter. However, the actual cfra is unknown when loading movies or sequences, so the iuser->framenr value itself was passed in its place, leading to incremental addition of the iuser frame offset. Removed the cfra parameter altogether from the image path function. This should instead be done separately if necessary, it's not an inherent part of constructing the image file name.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@47657 954f8c5b-7b00-dc11-b283-0030488c597c