Fix part 2 of [#31840] Quick Explode Bugs related to it's fade option.
This in fact had nothing to see with this operator or its fade option, reports were simply not shown when repeating an operator.

Carefully checked/tracked all calls to wm_operator_exec and WM_operator_repeat, and could not see any reason why this was this way, so I disabled this. Very easy to undo in case ugly problems arize (but I really do not expect any!).

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@48085 954f8c5b-7b00-dc11-b283-0030488c597c