Fixes [#31433] BMesh: Knife tool Angle Constraint function
The problem was that calculating the angle was reliant on previous mouse coordinates, which did not update to match snapped-to vertices and edges.

This solves the issue by updating mouse coordinates in knife_find_closest_edge() and knife_find_closest_vertex(). Additionally, when angle constraints are enabled, edge/vertex snapping is now prevented.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@48155 954f8c5b-7b00-dc11-b283-0030488c597c